Move risco coordinator to separate module
