+++ homeassistant/components/risco/sensor.py
@@ -17,8 +17,9 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 from homeassistant.util import dt as dt_util
 
-from . import RiscoEventsDataUpdateCoordinator, is_local
+from . import is_local
 from .const import DOMAIN, EVENTS_COORDINATOR
+from .coordinator import RiscoEventsDataUpdateCoordinator
 from .entity import zone_unique_id
 
 CATEGORIES = {
